openvino: rename quant weight header
