+++ ggml/src/ggml-openvino/ggml-openvino-op-support.cpp
@@ -202,7 +202,7 @@ static bool is_op_unsupported_case(const ggml_tensor * op) {
         if (op->ne[0] == 256 && (op->src[0]->type == GGML_TYPE_Q4_K || op->src[0]->type == GGML_TYPE_Q5_K ||
                                  op->src[0]->type == GGML_TYPE_Q4_1 || op->src[0]->type == GGML_TYPE_Q5_1)) {
             // These are all f16-arithmetic dequant rounding errors that intermittently exceed the
-            // tight 1e-7 NMSE threshold depending on the random test data (see ggml-quants.cpp
+            // tight 1e-7 NMSE threshold depending on the random test data (see ggml-openvino-quant-weights.cpp
             // make_int8_weights/make_int4_weights: dequant is done in f16, not f32, to keep the
             // Convert/Subtract/Multiply chain fusable into GatherMatmulCompressed/FullyConnectedCompressed
             // for the shared non-test code paths).
