+++ ggml/src/ggml-openvino/ggml-openvino-quant-extract.cpp
@@ -1,4 +1,4 @@
-#include "ggml-quants.h"
+#include "ggml-openvino-quant-weights.h"
 
 #include <algorithm>
 #include <cmath>
